Split nav events from Navigator.kt
